+++ .idea/modules/Readio.main.iml
@@ -1,6 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <module version="4">
-  <component name="CheckStyle-IDEA-Module" serialisationVersion="2">
-    <option name="activeLocationsIds" />
+  <component name="AdditionalModuleElements">
+    <content url="file://$MODULE_DIR$/../../Readio/src/main" dumb="true">
+      <sourceFolder url="file://$MODULE_DIR$/../../Readio/src/main/resources" type="java-resource" />
+    </content>
   </component>
 </module>
