--- Readio/src/main/java/ninegle/Readio/admin/domain/Role.java
@@ -1,6 +1,0 @@
-package ninegle.Readio.admin.domain;
-
-public enum Role {
-    ADMIN, USER
-
-}
