--- Readio/src/main/java/ninegle/Readio/admin/dto/LoginRequestDto.java
@@ -1,13 +1,0 @@
-package ninegle.Readio.admin.dto;
-
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class LoginRequestDto {
-    private final String email;
-    private final String password;
-
-}
