--- Readio/src/main/java/ninegle/Readio/admin/dto/SingUpRequestDto.java
@@ -1,15 +1,0 @@
-package ninegle.Readio.admin.dto;
-
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class SingUpRequestDto {
-    private final String email;
-    private final String password;
-    private final String nickname;
-    private final String phoneNumber;
-    private final int point;  // 포인트 필드 추가
-}
