--- Readio/src/main/java/ninegle/Readio/admin/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package ninegle.Readio.admin.repository;
-
-import ninegle.Readio.admin.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository <User, Long> {
-
-    Optional<User> findByEmail(String email);
-}
