+++ Readio/src/main/java/ninegle/Readio/mypage/service/PointService.java
@@ -3,9 +3,9 @@ package ninegle.Readio.mypage.service;
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
-import ninegle.Readio.admin.domain.User;
-import ninegle.Readio.admin.repository.UserRepository;
-import ninegle.Readio.admin.service.UserContextService;
+import ninegle.Readio.user.domain.User;
+import ninegle.Readio.user.repository.UserRepository;
+import ninegle.Readio.user.service.UserContextService;
 import ninegle.Readio.global.exception.BusinessException;
 import ninegle.Readio.global.exception.domain.ErrorCode;
 
@@ -17,7 +17,7 @@ public class PointService {
 	private final UserRepository userRepository;
 
 	public long getUserPoints() {
-		Long userId = userContextService.getCurrentAdminId();
+		Long userId = userContextService.getCurrentUserId();
 
 		User user = userRepository.findById(userId)
 			.orElseThrow(() -> new BusinessException(ErrorCode.AUTHENTICATION_REQUIRED));
