+++ Readio/src/main/java/ninegle/Readio/user/domain/Role.java
@@ -1,0 +1,6 @@
+package ninegle.Readio.user.domain;
+
+public enum Role {
+	ADMIN, USER
+
+}
