+++ Readio/src/main/java/ninegle/Readio/user/domain/User.java
@@ -1,0 +1,61 @@
+package ninegle.Readio.user.domain;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class User {
+
+	@Id
+	@Column(name = "user_id")
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+
+	@Column(length = 40, nullable = false)
+	private String email;
+
+	@Column(length = 100, nullable = false)
+	private String password;
+
+	private Role role = Role.USER;
+
+	private String nickname;
+
+	private String phoneNumber;
+
+	private long point = 50000;
+
+	@Builder
+	public User(String email, String password, String nickname, String phoneNumber, long point) {
+		this.email = email;
+		this.password = password;
+		this.nickname = nickname;
+		this.phoneNumber = phoneNumber;
+		this.point = point;
+	}
+
+	// 닉네임 변경 메서드
+	public void updateNickname(String nickname) {
+		this.nickname = nickname;
+	}
+
+	// 핸드폰번호 변경 메서드
+	public void updatePhoneNumber(String phone_number) {
+		this.phoneNumber = phone_number;
+	}
+
+	// point 수정할 수 있도록 setter 추가
+	public void setPoint(Long point) {
+		this.point = point;
+	}
+
+}
