+++ Readio/src/main/java/ninegle/Readio/user/dto/LoginRequestDto.java
@@ -1,0 +1,12 @@
+package ninegle.Readio.user.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class LoginRequestDto {
+	private String email;
+	private String password;
+
+}
