+++ Readio/src/main/java/ninegle/Readio/user/dto/RefreshTokenRequestDto.java
@@ -1,17 +1,13 @@
-package ninegle.Readio.admin.dto;
-
+package ninegle.Readio.user.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
+
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-
 @Getter
-@AllArgsConstructor
 @NoArgsConstructor
-
 public class RefreshTokenRequestDto {
-    @JsonProperty("refreshToken")
-    private String refreshToken;
+	@JsonProperty("refreshToken")
+	private String refreshToken;
 }
