+++ Readio/src/main/java/ninegle/Readio/user/dto/SingUpRequestDto.java
@@ -1,0 +1,14 @@
+package ninegle.Readio.user.dto;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class SingUpRequestDto {
+	private final String email;
+	private final String password;
+	private final String nickname;
+	private final String phoneNumber;
+	private long point = 50000L;
+}
