+++ Readio/src/main/java/ninegle/Readio/user/repository/UserRepository.java
@@ -1,0 +1,14 @@
+package ninegle.Readio.user.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import ninegle.Readio.user.domain.User;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+	Optional<User> findByEmail(String email);
+}
